requirements/base.txt
# base dependencies

Django
django-allauth
django-rest-auth
django-rest-swagger
django-widget-tweaks
djangorestframework
#mysqlclient
Pillow
django-ckeditor==5.4.0
django-cleanup
django-jet
Fabric3==1.13.1.post1
git+https://github.com/akoikelov/djazz
pypugjs
psycopg2
Celery
redis
django-filter
django-cors-headers